Fix macOS build: resolve ambiguous operator and type errors on ARM
On ARM macOS, size_t is unsigned long while uint64_t is unsigned long long
(both 64-bit but distinct types), causing several build failures:

- Ambiguous operator< between static_val<uint64_t> and size_type: add
  explicit static_cast<uint64_t>() on .size() calls in for-loop bounds
- std::clamp/std::max type deduction failures: use static_cast<uint64_t>()
  or explicit uint64_t type instead of UL suffix literals
- ConfigParameter<size_t> not in variant type list: change to uint64_t
- val<size_t&> not convertible to val<uint64_t>: change member type to
  uint64_t and read via getMemberWithOffset<uint64_t>
- DOMAIN macro conflict with macOS math.h: rename to SOCKET_DOMAIN

+++ nes-input-formatters/include/InputFormatters/FieldOffsets.hpp
@@ -71,7 +71,7 @@ class FieldOffsets final : public FieldIndexFunction<FieldOffsets<NumOffsetsPerF
     applyHasNext(const nautilus::val<uint64_t>& recordIdx, nautilus::val<FieldOffsets*> fieldOffsetsPtr) const
     {
         nautilus::val<uint64_t> totalNumberOfTuples
-            = *getMemberWithOffset<size_t>(fieldOffsetsPtr, offsetof(FieldOffsets, totalNumberOfTuples));
+            = *getMemberWithOffset<uint64_t>(fieldOffsetsPtr, offsetof(FieldOffsets, totalNumberOfTuples));
         return recordIdx < totalNumberOfTuples;
     }
 
@@ -225,7 +225,7 @@ private:
     FieldIndex sizeOfFieldDelimiter{};
     size_t numberOfFieldsInSchema{};
     size_t numberOfOffsetsPerTuple{};
-    size_t totalNumberOfTuples{};
+    uint64_t totalNumberOfTuples{};
     FieldIndex offsetOfFirstTuple{};
     FieldIndex offsetOfLastTuple{};
     std::vector<FieldIndex> indexValues;
