+++ nes-nautilus/src/Nautilus/Interface/BufferRef/RowTupleBufferRef.cpp
@@ -52,7 +52,7 @@ Record RowTupleBufferRef::readRecord(
     Record record;
     const auto bufferAddress = recordBuffer.getMemArea();
     const auto recordOffset = bufferAddress + (tupleSize * recordIndex);
-    for (nautilus::static_val<uint64_t> i = 0; i < fields.size(); ++i)
+    for (nautilus::static_val<uint64_t> i = 0; i < static_cast<uint64_t>(fields.size()); ++i)
     {
         const auto& [name, type, fieldOffset] = fields.at(i);
         if (not includesField(projections, name))
@@ -74,7 +74,7 @@ void RowTupleBufferRef::writeRecord(
 {
     const auto bufferAddress = recordBuffer.getMemArea();
     const auto recordOffset = bufferAddress + (tupleSize * recordIndex);
-    for (nautilus::static_val<uint64_t> i = 0; i < fields.size(); ++i)
+    for (nautilus::static_val<uint64_t> i = 0; i < static_cast<uint64_t>(fields.size()); ++i)
     {
         const auto& [name, type, fieldOffset] = fields.at(i);
         if (not rec.hasField(name))
