+++ nes-physical-operators/src/Join/HashJoin/HJBuildPhysicalOperator.cpp
@@ -138,7 +138,7 @@ void HJBuildPhysicalOperator::execute(ExecutionContext& ctx, Record& record) con
         hashMapPtr, hashMapOptions.fieldKeys, hashMapOptions.fieldValues, hashMapOptions.entriesPerPage, hashMapOptions.entrySize};
 
     /// Calling the key functions to add/update the keys to the record
-    for (nautilus::static_val<uint64_t> i = 0; i < hashMapOptions.fieldKeys.size(); ++i)
+    for (nautilus::static_val<uint64_t> i = 0; i < static_cast<uint64_t>(hashMapOptions.fieldKeys.size()); ++i)
     {
         const auto& [fieldIdentifier, type, fieldOffset] = hashMapOptions.fieldKeys[i];
         const auto& function = hashMapOptions.keyFunctions[i];
